filed: move reportcmd to its own file
