+++ core/src/filed/CMakeLists.txt
@@ -48,6 +48,7 @@ set(FDSRCS
     filed_conf.cc
     restore.cc
     status.cc
+    report.cc
     filed_utils.cc
 )
